# api/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise Fleet Management Platform API
  version: 1.0.0-locked
  description: >-
    Locked API contract for the Enterprise Fleet Management Platform.
    This document is generated from docs/architecture/00-locked-decisions.md and
    the PostgreSQL schema in db/schema. The frontend team binds the externally
    supplied UI designs to these endpoints (decision A3.6). No UI prescription
    is contained here.

    Conventions (decisions D1, D4, D7, C5.1):
      * Base path /api/v1. JSON only.
      * All timestamps are RFC 3339 in UTC (stored as timestamptz, rendered EAT client-side per A2.3).
      * Cursor pagination via `cursor` / `limit` query params; cursors are opaque.
      * Errors use RFC 7807 problem+json with an `error_code` extension member.
      * Every state-changing request MUST carry `Idempotency-Key: <uuid>` (C5.1).
  contact:
    name: Fleet Platform Architecture
servers:
  - url: https://api.fleet.internal/api/v1
    description: Production (AWS af-south-1, decision A1.10)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    idempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
  parameters:
...
    analyticsFrom:
      name: from
      in: query
      schema: { type: string, format: date-time }
      description: Start of the analytics window. Defaults to 30 days before `to`.
    analyticsTo:
      name: to
      in: query
      schema: { type: string, format: date-time }
      description: End of the analytics window. Defaults to now.
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem+json.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, example: https://docs.fleet.internal/problems/clockout_pending }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uuid }
        error_code:
          type: string
          description: Stable machine-readable code (see Error Codes).
          example: CLOCKOUT_PENDING
        field_errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...
    AnalyticsKpis:
      type: object
      description: Headline counters, already narrowed to the caller's resolved scope.
      required: [vehicles, drivers, distanceKm, fuelCost, anomalies]
      properties:
        vehicles: { type: number }
        drivers: { type: number }
        distanceKm: { type: number }
        fuelCost: { type: number }
        anomalies: { type: number }
    VehicleAnalytics:
      type: object
      required: [vehicle_id, plate, distanceKm, fuelCost, utilisationPct, anomalies]
      properties:
        vehicle_id: { type: string, format: uuid }
        plate: { type: string, nullable: true }
        distanceKm: { type: number }
        fuelCost: { type: number }
        utilisationPct: { type: number, description: Share of days in range with at least one shift. }
        anomalies: { type: number }
    DriverAnalytics:
      type: object
      required: [driver_id, name, distanceKm, shifts, anomalies]
      properties:
        driver_id: { type: string, format: uuid }
        name: { type: string, nullable: true }
        distanceKm: { type: number }
        shifts: { type: number }
        anomalies: { type: number }
    ManagerAnalyticsSummary:
      type: object
      description: One FLEET_MANAGER row inside the company roll-up, scored against their own scope.
      required: [user_id, email, assignedVehicleIds, assignedDriverIds, kpis]
      properties:
        user_id: { type: string, format: uuid }
        full_name: { type: string, nullable: true }
        email: { type: string }
        assignedVehicleIds:
          type: array
          nullable: true
          description: null = unrestricted on this dimension; [] = nothing assigned.
          items: { type: string, format: uuid }
        assignedDriverIds:
          type: array
          nullable: true
          items: { type: string, format: uuid }
        kpis: { $ref: '#/components/schemas/AnalyticsKpis' }
    CompanyAnalytics:
      type: object
      description: >
        Company roll-up plus one row per FLEET_MANAGER. The flat counters (active_fleet …
        anomalies_open) mirror the mobile AnalyticsReportSchema so GET /reports/analytics parses
        this body unchanged.
      required: [tenant_id, from, to, kpis, managers, active_fleet, open_accidents, pending_dvir, expiring_docs, fuel_spend_30d, anomalies_open]
      properties:
        tenant_id: { type: string, format: uuid }
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        kpis: { $ref: '#/components/schemas/AnalyticsKpis' }
        managers:
          type: array
          items: { $ref: '#/components/schemas/ManagerAnalyticsSummary' }
        active_fleet: { type: number }
        open_accidents: { type: number }
        pending_dvir: { type: number }
        expiring_docs: { type: number }
        fuel_spend_30d: { type: number }
        anomalies_open: { type: number }
    ManagerAnalytics:
      type: object
      description: One manager's slice, expanded per vehicle and per driver.
      required: [user_id, email, from, to, assignedVehicleIds, assignedDriverIds, kpis, vehicles, drivers]
      properties:
        user_id: { type: string, format: uuid }
        full_name: { type: string, nullable: true }
        email: { type: string }
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        assignedVehicleIds:
          type: array
          nullable: true
          items: { type: string, format: uuid }
        assignedDriverIds:
          type: array
          nullable: true
          items: { type: string, format: uuid }
        kpis: { $ref: '#/components/schemas/AnalyticsKpis' }
        vehicles:
          type: array
          items: { $ref: '#/components/schemas/VehicleAnalytics' }
        drivers:
          type: array
          items: { $ref: '#/components/schemas/DriverAnalytics' }
    VehicleAnalyticsDetail:
      type: object
      required: [from, to, vehicle, kpis]
      properties:
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        vehicle: { $ref: '#/components/schemas/VehicleAnalytics' }
        kpis: { $ref: '#/components/schemas/AnalyticsKpis' }
    DriverAnalyticsDetail:
      type: object
      required: [from, to, driver, kpis]
      properties:
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        driver: { $ref: '#/components/schemas/DriverAnalytics' }
        kpis: { $ref: '#/components/schemas/AnalyticsKpis' }
...
  responses:
...
    Unauthorized:
      description: Missing / expired token.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Forbidden:
      description: Authenticated but lacking permission.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    NotFound:
      description: Resource not found.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
paths:
...
  /analytics/company:
    get:
      tags: [Analytics]
      summary: Company KPIs plus a per-manager breakdown.
      description: >
        Also served at GET /reports/analytics. The flat AnalyticsReport counters are included in the
        body, so the mobile AnalyticsReportSchema parses this response unchanged.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/analyticsFrom'
        - $ref: '#/components/parameters/analyticsTo'
      responses:
        '200':
          description: Company roll-up scoped to the caller.
          content: { application/json: { schema: { $ref: '#/components/schemas/CompanyAnalytics' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /analytics/manager/{userId}:
    get:
      tags: [Analytics]
      summary: One manager's KPIs, expanded per vehicle and per driver.
      description: Authorised for an ADMIN of the tenant, or for that manager themselves.
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/analyticsFrom'
        - $ref: '#/components/parameters/analyticsTo'
      responses:
        '200':
          description: Manager analytics.
          content: { application/json: { schema: { $ref: '#/components/schemas/ManagerAnalytics' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /analytics/vehicle/{vehicleId}:
    get:
      tags: [Analytics]
      summary: Per-vehicle KPIs (tenant- and scope-checked).
      security:
        - bearerAuth: []
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/analyticsFrom'
        - $ref: '#/components/parameters/analyticsTo'
      responses:
        '200':
          description: Vehicle analytics.
          content: { application/json: { schema: { $ref: '#/components/schemas/VehicleAnalyticsDetail' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /analytics/driver/{driverId}:
    get:
      tags: [Analytics]
      summary: Per-driver KPIs. Authorised for the driver, their manager, or an ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - name: driverId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/analyticsFrom'
        - $ref: '#/components/parameters/analyticsTo'
      responses:
        '200':
          description: Driver analytics.
          content: { application/json: { schema: { $ref: '#/components/schemas/DriverAnalyticsDetail' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /analytics/me:
    get:
      tags: [Analytics]
      summary: The signed-in driver's own KPIs.
      description: >
        Returns DriverAnalyticsDetail for a DRIVER. A caller with no driver row (ADMIN, manager)
        receives the CompanyAnalytics view of their own scope instead.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/analyticsFrom'
        - $ref: '#/components/parameters/analyticsTo'
      responses:
        '200':
          description: The caller's own analytics.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/DriverAnalyticsDetail'
                  - $ref: '#/components/schemas/CompanyAnalytics'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }